# docs/openapi.yaml
openapi: 3.1.0
info:
  title: APEX Recommendation API
  description: 'Production-grade AI recommendation engine with a 6-model ensemble
    (SASRec, KAN, LightGCN, Diffusion, Quantum-Fluid, Hyperbolic). Implements Doubly
    Robust IPS weight optimization, differential privacy, and adaptive 3-tier serving
    (GPU / ONNX CPU / FAISS lite).


    **Interactive docs:** `/docs` (Swagger UI) · `/redoc` (ReDoc)


    **Full API reference:** [docs/API_REFERENCE.md](https://github.com/pavanpajjuri/Movie-Recommendation-System/blob/main/docs/API_REFERENCE.md)'
  contact:
    name: APEX Project
    url: https://github.com/pavanpajjuri/Movie-Recommendation-System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.0
paths:
...
  /v1/evaluation/recommendations:
    get:
      tags:
      - Evaluation
      summary: Recommendation Quality Report
      operationId: recommendation_quality_report_v1_evaluation_recommendations_get
      parameters:
      - name: sample_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 25
          title: Sample Size
      - name: k
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 10
          title: K
      - name: X-Nova-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Nova-Api-Key
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Id
      - name: X-Catalog-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Catalog-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/evaluation/semantic-benchmark:
    get:
      tags:
      - Evaluation
      summary: Semantic Benchmark Report
      operationId: semantic_benchmark_report_v1_evaluation_semantic_benchmark_get
      parameters:
      - name: k
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 10
          title: K
      - name: sync
        in: query
        required: false
        schema:
          type: boolean
          description: Compute synchronously instead of returning async cache warming
            status
          default: false
          title: Sync
        description: Compute synchronously instead of returning async cache warming
          status
      - name: X-Nova-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Nova-Api-Key
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Id
      - name: X-Catalog-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Catalog-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/evaluation/search-benchmark:
    get:
      tags:
      - Evaluation
      summary: Search Benchmark Report
      operationId: search_benchmark_report_v1_evaluation_search_benchmark_get
      parameters:
      - name: k
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 1
          default: 5
          title: K
      - name: X-Nova-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Nova-Api-Key
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Id
      - name: X-Catalog-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Catalog-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/evaluation/recommendation-benchmark:
    get:
      tags:
      - Evaluation
      summary: Recommendation Benchmark Report
      operationId: recommendation_benchmark_report_v1_evaluation_recommendation_benchmark_get
      parameters:
      - name: k
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 10
          title: K
      - name: sync
        in: query
        required: false
        schema:
          type: boolean
          description: Compute synchronously instead of returning async cache warming
            status
          default: false
          title: Sync
        description: Compute synchronously instead of returning async cache warming
          status
      - name: X-Nova-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Nova-Api-Key
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Id
      - name: X-Catalog-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Catalog-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
servers:
- url: http://localhost:8000
  description: Local development
- url: https://your-api.onrender.com
  description: Production (Render free tier — Tier 3)